bot/requirements.txt
python-telegram-bot==21.4
Pillow
numpy
